BookDetails implemented, checkers implemented, loans changed, slight fixes

+++ NTIlliaProject/src/main/java/com/illia/project/ntilliaproject/controller/dto/review/CreateReviewDto.java
@@ -4,24 +4,24 @@ import java.util.Date;
 
 public class CreateReviewDto {
 
-    private int rating;
+    private double rating;
 
     private String comment;
 
     private Date reviewDate;
 
-    public CreateReviewDto( int rating, String comment, Date reviewDate) {
+    public CreateReviewDto( double rating, String comment, Date reviewDate) {
         this.rating = rating;
         this.comment = comment;
         this.reviewDate = reviewDate;
     }
 
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
