+++ NTIlliaProject/src/main/java/com/illia/project/ntilliaproject/controller/dto/review/CreateReviewResponseDto.java
@@ -5,13 +5,13 @@ import java.util.Date;
 public class CreateReviewResponseDto {
     private int reviewID;
 
-    private int rating;
+    private double rating;
 
     private String comment;
 
     private Date reviewDate;
 
-    public CreateReviewResponseDto(int reviewID, int rating, String comment, Date reviewDate) {
+    public CreateReviewResponseDto(int reviewID, double rating, String comment, Date reviewDate) {
         this.reviewID = reviewID;
         this.rating = rating;
         this.comment = comment;
@@ -26,11 +26,11 @@ public class CreateReviewResponseDto {
         this.reviewID = reviewID;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
